+++ sentry-android-core/src/main/java/io/sentry/android/core/IDebugImagesLoader.java
@@ -1,0 +1,13 @@
+package io.sentry.android.core;
+
+import io.sentry.protocol.DebugImage;
+import java.util.List;
+import org.jetbrains.annotations.ApiStatus;
+
+/** Used for loading the list of debug images from sentry-native. */
+@ApiStatus.Internal
+public interface IDebugImagesLoader {
+  List<DebugImage> loadDebugImages();
+
+  void clearDebugImages();
+}